Move `FormTextField` to new directory `common/components/form`
